Former-commit-id: 218d5be11cf1ea6d7f6a4649aa34e97adcb0d468
Former-commit-id: 10f7c785ab33b3d5b08894c21b426dfe4acd81df

+++ registration-processor/registration-processor-core/src/main/java/io/mosip/registration/processor/core/abstractverticle/StageHealthCheckHandler.java
@@ -185,17 +185,19 @@ public class StageHealthCheckHandler implements HealthCheckHandler {
 			System.setProperty("hadoop.home.dir", hadoopLibPath.toString());
 			if (SystemUtils.IS_OS_WINDOWS) {
 				File binPath = FileUtils.getFile(hadoopLibPath.toString(), "bin");
-				binPath.mkdir();
+				boolean created=binPath.mkdir();
 				Resource resource = resourceLoader.getResource(CLASSPATH_PREFIX + WIN_UTIL);
 				if (resource.exists()) {
-					java.nio.file.Path c = Paths.get("s", "x");
+
+					if(created) {
 					File winUtilsPath = FileUtils.getFile(binPath.toString(), resource.getFilename());
-					FileUtils.copyInputStreamToFile(resource.getInputStream(), winUtilsPath);
-				}
+					   FileUtils.copyInputStreamToFile(resource.getInputStream(), winUtilsPath);
+					}
+					}
 			}
 
 			if (isAuthEnable) {
-				configuration = initSecurityConfiguration(configuration);
+				initSecurityConfiguration(configuration);
 				configuredFileSystem = FileSystem.get(configuration);
 			} else {
 				configuredFileSystem = getDefaultConfiguredFileSystem(configuration);
@@ -233,7 +235,7 @@ public class StageHealthCheckHandler implements HealthCheckHandler {
 	 * @return
 	 * @throws Exception
 	 */
-	private Configuration initSecurityConfiguration(Configuration configuration) throws Exception {
+	private void initSecurityConfiguration(Configuration configuration) throws Exception {
 		configuration.set("dfs.data.transfer.protection", "authentication");
 		configuration.set("hadoop.security.authentication", "kerberos");
 		InputStream krbStream = getClass().getClassLoader().getResourceAsStream("krb5.conf");
@@ -243,7 +245,6 @@ public class StageHealthCheckHandler implements HealthCheckHandler {
 		UserGroupInformation.setConfiguration(configuration);
 		String user = hdfsUserName + "@" + kdcDomain;
 		loginWithKeyTab(user, keytabPath);
-		return configuration;
 	}
 
 	/**
@@ -255,8 +256,8 @@ public class StageHealthCheckHandler implements HealthCheckHandler {
 		File keyPath = null;
 		Resource resource = resourceLoader.getResource(keytabPath);
 		File dataPath = FileUtils.getFile(hadoopLibPath.toString(), "data");
-		dataPath.mkdir();
-		if (resource.exists()) {
+		boolean created=dataPath.mkdir();
+		if (resource.exists() && created) {
 			keyPath = FileUtils.getFile(dataPath.toString(), resource.getFilename());
 			FileUtils.copyInputStreamToFile(resource.getInputStream(), keyPath);
 		} else {
